Drop special handling for UTF-16.
Even the CXML maintainer admits it doesn't really work. From a
discussion in babel-devel in 2009:

"As a side note, I find it a huge hassle to write code portable between
the Lisp implementations with Unicode support. For CXML, I needed read
time conditionals checking for UTF-16 Lisps. And it still doesn't
actually work, because most of the other free libraries like Babel,
CL-Unicode, and in turn, CL-PPCRE, expect 21 bit characters and are
effectively broken on Allegro and LispWorks."